Remove ~/.gitignore as the global ignoring file list.
Because specifying ~/.gitignore as the global ignoring list,
in some ther project we do not `git add` public directory.

This behaviour is not out intention.
And it seems that the reason for selecting ~/.gitignore
as the global ignotring file list is groundless.

So, we remove ~/.gitignore as the global ignoring file list.

This commit fixed #269.